aws/request: Simplify connection reset detection
Simplifies the SDK's connection reset detection, and removes the usage
of deprecated syscall package.

Fix #1263
