--- aws/request/connection_reset_error_other_test.go
@@ -1,12 +1,0 @@
-// +build appengine plan9
-
-package request_test
-
-import (
-	"errors"
-)
-
-var (
-	errAcceptConnectionResetStub = errors.New("accept: connection reset")
-	errReadConnectionResetStub   = errors.New("read: connection reset")
-)
